services/user_routing_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pymongo==4.8.0
httpx==0.27.0
python-dotenv==1.0.1